More tidy-up

+++ libraries/KustoLoco.Core/Evaluation/BuiltIns/Aggregates/MinIfAggregate.cs
@@ -29,7 +29,7 @@ internal partial class MinIfAggregate
     }
 
     internal static long? LongImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : (long)context.LongValue;
+        => context.Count == 0 ? null : context.LongValue;
 
     internal static double DoubleImpl(NumericAggregate context, double n, bool t)
     {
@@ -66,7 +66,7 @@ internal partial class MinIfAggregate
     }
 
     internal static TimeSpan? TsImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : new TimeSpan((long)context.LongValue);
+        => context.Count == 0 ? null : new TimeSpan(context.LongValue);
 
     internal static DateTime DtImpl(NumericAggregate context, TimeSpan n, bool t)
     {
@@ -78,5 +78,5 @@ internal partial class MinIfAggregate
     }
 
     internal static DateTime? DtImplFinish(NumericAggregate context)
-        => context.Count == 0 ? null : new DateTime((long)context.LongValue, DateTimeKind.Utc);
+        => context.Count == 0 ? null : new DateTime(context.LongValue, DateTimeKind.Utc);
 }
